updated snapshot, renaming of ProjectsPage
