+++ app/src/features/projects/ProjectsListPage.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
 import { render } from '@testing-library/react';
-import ProjectsPage from './ProjectsPage';
+import ProjectsPage from './ProjectsListPage';
 
 // https://stackoverflow.com/questions/58524183/how-to-mock-history-push-with-the-new-react-router-hooks-using-jest/59451956
